ggml-cpu: move MUL_MAT_ID_COLD kernel to its own file
- add ggml-cpu-mul-mat-id-cold.c with the cold-expert kernel
- add shared header for mmid helpers (mmid_row_mapping, incr_ptr_aligned, one_chunk)
- drop static on the two helpers now shared with stock mul_mat_id
